refactor grpc controller, jwt util
