+++ pkg/proto/product.proto
@@ -1,0 +1,6 @@
+syntax="proto3";
+
+option go_package = "./pkg/pb/product";
+
+service ProductService {
+}
